+++ client/src/components/Cell.js
@@ -1,7 +1,8 @@
 import { React, useState, useEffect } from 'react';
 import { Hexagon, Text } from 'react-hexgrid'
 import { Player } from './Player' // Import the Player component
-import { HandleClick, MovePlayer1, MovePlayer2, MovePlayer } from './HandleClick'
+import { HandleClick } from './HandleClick'
+import { MovePlayer1, MovePlayer2, MovePlayer } from './PlayerHandleClick'
 import { PlayerSpawn1, PlayerSpawn2 } from './PlayerSpawn'
 // import { AdjacentTiles } from './AdjacentTiles';
 // import { movePlayerToRandomAdjacentTile, movePlayerToAdjacentTile } from './PlayerMovement';
@@ -32,7 +33,7 @@ const Cell = ({ q, r, i }) => {
 
             // MovePlayer(2, q, r, i, setClickedPlayer, playerLocations, setPlayerLocations, randomPlace1, randomPlace2);
 
-        }} key={i} q={q} r={r}>
+        }} key={`${i}-${q}-${r}`} q={q} r={r}>
             {<Text>{i} {q} {r}</Text>}
             {playerLocations.player1 === i ? <Player playerNumber={1} /> : null}
             {playerLocations.player2 === i ? <Player playerNumber={2} /> : null}
